Fixing #202 . Doubles the memory requirement in alignments

+++ biojava3-alignment/src/test/java/org/biojava3/alignment/NeedlemanWunschTest.java
@@ -101,7 +101,7 @@ public class NeedlemanWunschTest {
 
     @Test
     public void testGetScoreMatrix() {
-        short[][][] scores = alignment.getScoreMatrix();
+        int[][][] scores = alignment.getScoreMatrix();
         assertEquals(scores[2][1][0], -6);
         scores = self.getScoreMatrix();
         assertEquals(scores[3][4][2], 4);
